pyproject.toml
[project]
name = "NoteAI"
version = "0.1.0"
description = "AI驱动的Markdown笔记知识库管理桌面应用"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "NoteAI Contributors" }
]
dependencies = [
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.0",
    "tiktoken>=0.5.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.6",
    "readability-lxml>=0.8.1",
    "validators>=0.22.0",
    "pydantic>=2.5.0",
    "PyMuPDF>=1.23.0",
    "mammoth>=1.0.0",
    "python-docx>=1.0.0",
    "pillow>=10.0.0",
    "markdown>=3.5.0",
    "PyYAML>=6.0",
    "jieba>=0.42.1",
    "html2text>=2024.2.26",
    "watchdog>=4.0.0",
    "send2trash>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "modules*", "prompts*", "utils*", "tests*"]

# Maps import name `sidecar` → directory python/sidecar (Tauri sidecar package).
[tool.setuptools.package-dir]
"sidecar" = "python/sidecar"

[project.scripts]
noteai = "run:main"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = [".", "python"]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
